chore: move TRIPS data into server module and pass it through loader to simulate server API request
