+++ app/containers/index/trip-timeline/trip-timeline.tsx
@@ -2,7 +2,8 @@ import type { FC } from 'react';
 import { trim } from '~/utils/trim';
 import { Button } from '~/components/button';
 import { TripTimelineNode } from './trip-timeline-node';
-import { TRIPS } from '~/data-access/trips';
+import { useLoaderData } from 'react-router';
+import type { HomePageLoaderData } from '~/routes/index';
 
 interface TripTimelineProps {
     className?: string;
@@ -13,6 +14,8 @@ interface TripTimelineProps {
  *  Checkout issue: https://github.com/Alexius-Huang/travel.alexius-huang.dev/issues/24
  */
 export const TripTimeline: FC<TripTimelineProps> = ({ className }) => {
+    const { trips } = useLoaderData<HomePageLoaderData>();
+
     return (
         <section className={className}>
             <h2
@@ -36,7 +39,7 @@ export const TripTimeline: FC<TripTimelineProps> = ({ className }) => {
                     aria-hidden="true"
                 />
 
-                {TRIPS.map(trip => (
+                {trips.map(trip => (
                     <TripTimelineNode key={trip.id} tripDetails={trip} />
                 ))}
